+++ src/JsonApi/Response/JsonApiResponse.php
@@ -5,9 +5,9 @@ namespace WoohooLabs\Yang\JsonApi\Response;
 
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\StreamInterface;
-use WoohooLabs\Yang\JsonApi\Deserializer\DeserializerInterface;
-use WoohooLabs\Yang\JsonApi\Deserializer\JsonDeserializer;
 use WoohooLabs\Yang\JsonApi\Schema\Document;
+use WoohooLabs\Yang\JsonApi\Serializer\DeserializerInterface;
+use WoohooLabs\Yang\JsonApi\Serializer\JsonDeserializer;
 
 class JsonApiResponse implements ResponseInterface
 {