+++ src/JsonApi/Serializer/DeserializerInterface.php
@@ -1,14 +1,14 @@
 <?php
 declare(strict_types=1);
 
-namespace WoohooLabs\Yang\JsonApi\Deserializer;
+namespace WoohooLabs\Yang\JsonApi\Serializer;
 
 use Psr\Http\Message\ResponseInterface;
 
 interface DeserializerInterface
 {
     /**
-     * @return array|null
+     * @return array|mixed|null
      */
     public function deserialize(ResponseInterface $response);
 }